BFF 1033 Switch Overview page to stateless widget
